test(ui): isolate catalog activity coverage
